Move articles into src and configure markdown
