+++ decision-tree-app/package.json
@@ -28,6 +28,7 @@
     "eslint-plugin-react-refresh": "^0.4.20",
     "gh-pages": "^6.3.0",
     "globals": "^16.3.0",
-    "vite": "^7.0.4"
+    "vite": "^7.0.4",
+    "vite-plugin-md": "^0.22.5"
   }
 }
