+++ decision-tree-app/vite.config.js
@@ -1,7 +1,8 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import markdown from 'vite-plugin-md';
 
 export default defineConfig({
-  plugins: [react()],
+  plugins: [react(), markdown()],
   base: '/parsanaenergy/',
 });
